+++ LookupAnything/release-notes.md
@@ -1,7 +1,8 @@
 [← back to readme](README.md)
 
 # Release notes
-## Upcoming version
+## Upcoming release
+* Improved eradication goal info in monster lookups.
 * Internal refactor to translation handling.
 * Improved translations. Thanks to Enaium (updated Chinese), Macskasajt05 (added Hungarian), and zhxxn (updated Korean)!